only creating backup of zsh original files if they exist

+++ zsh/install.sh
@@ -29,8 +29,15 @@ git clone https://github.com/zsh-users/zsh-autosuggestions ${zsh_path}/plugins/z
 
 echo "🔧 Configuring Zsh..."
 
-mv ${HOME}/.zshrc ${HOME}/.zshrc.bkp
-mv ${HOME}/.zsh_history ${HOME}/.zsh_history.bkp
+if [ -f ${HOME}/.zshrc ]; then
+  echo "📁 .zshrc file already exists. Creating a backup..."
+  mv ${HOME}/.zshrc ${HOME}/.zshrc.bkp
+fi
+
+if [ -f ${HOME}/.zsh_history ]; then
+  echo "📁 .zsh_history file already exists. Creating a backup..."
+  mv ${HOME}/.zsh_history ${HOME}/.zsh_history.bkp
+fi
 
 files_to_link=(".zshrc", ".zsh_history")
